to-program/requirements.txt
autopep8==1.6.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
distlib==0.3.4
filelock==3.6.0
Flask==2.0.3
google-api-core==2.7.1
google-auth==2.6.2
google-cloud==0.34.0
google-cloud-core==2.2.3
google-cloud-datastore==2.4.0
google-cloud-firestore==2.4.0
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
lark-parser==0.12.0
libcst==0.4.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
platformdirs==2.5.1
plyer==2.0.0
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
toml==0.10.2
typing-inspect==0.7.1
typing_extensions==4.1.1
urllib3==1.26.9
virtualenv==20.14.0
Werkzeug==2.0.3
